+++ packages/@expo/cli/src/start/server/type-generation/routes.ts
@@ -305,6 +305,10 @@ const routerDotTSTemplate = unsafeTemplate`declare module "expo-router" {
     ? { params: RouteParams<InferPathName<T>> }
     : unknown;
 
+  /** Returns the search parameters for a route **/
+  export type SearchParams<T extends DynamicRouteTemplate | StaticRoutes | RelativePathString> =
+    T extends DynamicRouteTemplate ? RouteParams<T> : {};
+
   type InferPathName<T> = T extends { pathname: infer P } ? P : never;
 
   export interface LinkProps<T> extends OriginalLinkProps {
@@ -331,5 +335,6 @@ const routerDotTSTemplate = unsafeTemplate`declare module "expo-router" {
   };
 
   export function useRouter<T>(): Router<T>
+  export function useLocalSearchParams<T extends DynamicRouteTemplate | StaticRoutes | RelativePathString>(): SearchParams<T>
 }
 `;